+++ frontend/wasm/src/lib.rs
@@ -27,6 +27,7 @@ use alloc::rc::Rc;
 
 use component::build_ir::translate_component;
 use error::WasmResult;
+use midenc_frontend_wasm_metadata::PackageSections;
 use midenc_hir::{Context, dialects::builtin};
 use module::build_ir::translate_module_as_component;
 use wasmparser::WasmFeatures;
@@ -39,8 +40,8 @@ pub use self::{config::*, emit::WatEmit, error::WasmError};
 pub struct FrontendOutput {
     /// The IR component translated from the Wasm
     pub component: builtin::ComponentRef,
-    /// The serialized AccountComponentMetadata (name, description, storage layout, etc.)
-    pub account_component_metadata_bytes: Option<Vec<u8>>,
+    /// Metadata payloads to attach to the Miden package.
+    pub sections: PackageSections,
 }
 
 /// Translate a valid Wasm core module or Wasm Component Model binary into Miden
@@ -56,7 +57,7 @@ pub fn translate(
         let component = translate_module_as_component(wasm, config, context)?;
         Ok(FrontendOutput {
             component,
-            account_component_metadata_bytes: None,
+            sections: PackageSections::default(),
         })
     } else {
         translate_component(wasm, config, context)