+++ midenc-compile/src/pipeline/testing.rs
@@ -212,7 +212,10 @@ pub(crate) fn component_in_namespace(
     crate::MidenComponent {
         world,
         component: Some(component),
-        account_component_metadata_bytes: metadata,
+        sections: midenc_frontend_wasm_metadata::PackageSections {
+            account_component_metadata: metadata,
+            note_storage_schema: None,
+        },
         source_provenance: ProjectSourceProvenanceInputs {
             root: SourceFileProvenance {
                 path: FsPath::new(file!()).to_path_buf().into_boxed_path(),