+++ tests/integration-network/src/mockchain/notes/mod.rs
@@ -1,1 +1,2 @@
 mod basic_wallet;
+mod schema;